refactor(index): co-locate index/mod.rs strays into their duty modules
mod.rs had accumulated free functions and result types that belong with the code
that produces them. Move them to where they're owned, leaving mod.rs as a curated
index (405 -> 267 lines):

- mem_diag.rs: the opt-in memory diagnostics (mem_trace,
  maybe_set_sqlite_soft_heap_limit), re-exported so the crate::index::* call paths
  stay stable.
- ParserFailure -> parser_failures.rs, DiscoveryStatus -> discovery.rs (the
  methods that build them already live there).
- LogicalSymbolKey + LogicalSymbolMemberRow -> graph_index.rs (the logical-symbol
  rebuild owns them); file_index imports them for insert_logical_group.
- rename file_lifecycle.rs -> file_rows.rs to kill the near-homonym with
  lifecycle.rs.

No behavior change.
